Move action-queue to utils
