+++ src/app/inventory/move-item.ts
@@ -10,8 +10,8 @@ import _ from 'lodash';
 import { Subject } from 'rxjs';
 import { showNotification } from '../notifications/notifications';
 import { loadingTracker } from '../shell/loading-tracker';
+import { queueAction } from '../utils/action-queue';
 import { reportException } from '../utils/exceptions';
-import { queueAction } from './action-queue';
 import { updateCharacters } from './d2-stores';
 import { InventoryBucket } from './inventory-buckets';
 import { moveItemTo as moveTo } from './item-move-service';
